Warning, GIGANTIC COMMIT! Replaced Keyframe subclasses with Generic Types of Keyframe and replaced all instanceof calls Replaced implementations of Linear and Spline interpolation to interpolate any Object that extends KeyframeValue and contains at least one public double Field. I'll need this for Keyframe interpolation in CustomImageObject Transformations Made MarkerKeyframe *no* subclass of Keyframe to avoid conflicts with PositionKeyframe in instanceof checks for Keyframe#getValue Created KeyframeList which extends ArrayList to provide some helping functions which DRY up the ReplayHandler Split up ReplayHandler's keyframeList into timeKeyframeList and positionKeyframeList

+++ src/main/java/eu/crushedpixel/replaymod/entities/CameraEntity.java
@@ -160,8 +160,8 @@ public class CameraEntity extends EntityPlayer {
 
     public void moveAbsolute(Position pos) {
         this.moveAbsolute(pos.getX(), pos.getY(), pos.getZ());
-        rotationPitch = pos.getPitch();
-        rotationYaw = pos.getYaw();
+        rotationPitch = (float)pos.getPitch();
+        rotationYaw = (float)pos.getYaw();
     }
 
     public void moveAbsolute(double x, double y, double z) {
@@ -181,8 +181,8 @@ public class CameraEntity extends EntityPlayer {
     }
 
     public void movePath(Position pos) {
-        this.prevRotationPitch = this.rotationPitch = pos.getPitch();
-        this.prevRotationYaw = this.rotationYaw = pos.getYaw();
+        this.prevRotationPitch = this.rotationPitch = (float)pos.getPitch();
+        this.prevRotationYaw = this.rotationYaw = (float)pos.getYaw();
         this.lastTickPosX = this.prevPosX = this.posX = pos.getX();
         this.lastTickPosY = this.prevPosY = this.posY = pos.getY();
         this.lastTickPosZ = this.prevPosZ = this.posZ = pos.getZ();
